+++ tgbotapi.webapps/src/jsMain/kotlin/dev/inmo/tgbotapi/webapps/stories/StoryShareParams.kt
@@ -5,3 +5,13 @@ external interface StoryShareParams {
     @JsName("widget_link")
     val widgetLink: StoryWidgetLink?
 }
+
+fun StoryShareParams(
+    text: String,
+    widgetLink: StoryWidgetLink?
+): StoryShareParams {
+    val result: dynamic = js("{}")
+    result["text"] = text
+    widgetLink ?.let { result["widgetLink"] = it }
+    return result.unsafeCast<StoryShareParams>()
+}